Fix backend-core tests.
